Separated out the tests for ReadNameParser, added some negative tests for rapidParseInt and then fixed up the resulting failures.
